Moving to rank_transform module
