+++ python/ml4ir/applications/ranking/model/layers/__init__.py
@@ -1,3 +1,4 @@
 from ml4ir.applications.ranking.model.layers.set_rank_encoder import *
 from ml4ir.applications.ranking.model.layers.masking import *
 from ml4ir.applications.ranking.model.layers.normalization import *
+from ml4ir.applications.ranking.model.layers.rank_transform import *
